+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add ruby version to gem version output
 - Add ability last position in TUI selector
 - Add ability to isable readme and changelog viewing if they are unreachable
+- Add support for Codeberg readmes and changelogs (before only Github and Gitlab were supported)
 
 - Change pager to align TUI at the top of the screen
 - Change default pager to use less for consistent paging